pyproject.toml
[project]
name = "sqlgenerator"
version = "0.1.0"
description = "AI-powered SQL generator and data analyst (SQLite-based POC)"
requires-python = ">=3.11"

dependencies = [
    "pandas",
    "openpyxl",
    "tabulate",
    "numpy",
    "python-dotenv",
    "mcp[cli]",
]
# Note: sqlite3 is part of Python stdlib — no install needed
